Hammer properly works, Goron Mask drifting, Champion Tunic and Leathers added, cleaned up heavy boots code for smashing block, added majoras mask with logic and art, updated master ore, added master ore block, added secret stone block, added master shard, added refined master ore, recipe change for upgraded master sword 2/3, art for zora flippers

+++ src/main/java/com/kamth/zeldamod/item/armors/HeavyBoots.java
@@ -71,13 +71,12 @@ public class HeavyBoots extends ArmorItem {
         }
         {
             Level level = world;
-            if (player.isCrouching()){
-return;}
-            else if (level.getBlockState(player.getOnPos()).getMaterial() == Material.GLASS) {
+
+             if (level.getBlockState(player.getOnPos()).getMaterial() == Material.GLASS && !player.isCrouching()) {
                 level.destroyBlock(player.getOnPos(), false);}
-            if (player.isCrouching()){}
-            else if (level.getBlockState(player.getOnPos()).getBlock() == Blocks.ICE) {
-                level.destroyBlock(player.getOnPos(), true);
+
+             if (level.getBlockState(player.getOnPos()).getBlock() == Blocks.ICE && !player.isCrouching()) {
+                level.destroyBlock(player.getOnPos(), false);
                 return;}}
     }
